@@ -18,9 +18,20 @@
   },
   "license": "ISC",
   "author": "commenthol <commenthol@gmail.com>",
-  "main": "lib",
+  "maintainers": [
+    "commenthol <commenthol@gmail.com>"
+  ],
+  "type": "module",
+  "main": "./lib/index.cjs",
+  "exports": {
+    ".": {
+      "import": "./src/index.js",
+      "require": "./lib/index.cjs"
+    },
+    "./package.json": "./package.json"
+  },
   "bin": {
-    "holidays-ical": "bin/holidays-ical.js"
+    "holidays-ical": "bin/holidays-ical.cjs"
   },
   "directories": {
     "lib": "lib",
@@ -28,52 +39,33 @@
   },
   "scripts": {
     "all": "npm-run-all clean lint build test",
-    "build": "babel -d lib src",
+    "build": "rollup -c",
     "clean": "rimraf lib coverage .nyc_output",
     "clean:all": "npm-run-all clean clean:node_modules",
     "clean:node_modules": "rimraf node_modules",
-    "coverage": "nyc -r text -r html npm test",
-    "lint": "eslint '**/*.js'",
+    "coverage": "c8 -r text -r html npm run test:ci",
+    "lint": "eslint \"**/*.js\"",
     "prepublishOnly": "npm run all",
     "test": "mocha"
   },
-  "babel": {
-    "presets": [
-      "env"
-    ]
-  },
-  "eslintConfig": {
-    "plugins": [
-      "standard"
-    ],
-    "extends": "standard",
-    "rules": {
-      "no-console": 2
-    }
-  },
   "dependencies": {
-    "commander": "^5.0.0",
-    "date-holidays": "^1.5.0"
+    "commander": "^7.0.0",
+    "date-holidays": "^3.0.0-0"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-preset-env": "^1.7.0",
-    "eslint": "^6.8.0",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-plugin-import": "^2.20.2",
+    "c8": "^7.5.0",
+    "eslint": "^7.19.0",
+    "eslint-config-standard": "^16.0.2",
+    "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "mocha": "^7.1.1",
+    "eslint-plugin-promise": "^4.3.1",
+    "eslint-plugin-standard": "^5.0.0",
+    "mocha": "^8.3.0",
     "npm-run-all": "^4.1.5",
-    "nyc": "^15.0.1",
-    "rimraf": "^3.0.2"
+    "rimraf": "^3.0.2",
+    "rollup": "^2.38.5"
   },
   "engines": {
-    "node": ">=8.0.0"
-  },
-  "maintainers": [
-    "commenthol <commenthol@gmail.com>"
-  ]
+    "node": ">=12.0.0"
+  }
 }
