@@ -1,11 +1,11 @@
-const tVcalendar = (vevents) => `BEGIN:VCALENDAR
+export const tVcalendar = (vevents) => `BEGIN:VCALENDAR
 VERSION:2.0
 PRODID:-//date/holidays//NONSGML v1.0//EN
 METHOD:PUBLISH
 ${vevents.join('')}END:VCALENDAR
 `
 
-const tVevent = (event) => `BEGIN:VEVENT
+export const tVevent = (event) => `BEGIN:VEVENT
 CREATED:${event.created}
 LAST-MODIFIED:${event.created}
 DTSTAMP:${event.created}
@@ -16,8 +16,3 @@ TRANSP:${event.busy ? 'OPAQUE' : 'TRANSPARENT'}
 UID:${event.uid}
 END:VEVENT
 `
-
-module.exports = {
-  tVcalendar,
-  tVevent
-}
